Create README.md
I worked and currently working as a Virtual Assistant and Digital Marketing Admin with different tasks on it. Through the skills that I have and gained through education and past job experiences while on the other hand, it hones my skills and provides the additional field of expertise that can help me in the future. I decided to apply to another company because I am too fast to finish my tasks on time and I want more tasks. Feel free to contact me. My details are attached on my resume submitted.